refactor(market): clean up and re-organize the code
