+++ libs/solana-dapp/market/utils/get-prices/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/get-prices';
+export * from './lib/types';
