+++ libs/solana-dapp/market/utils/get-prices/src/lib/types.ts
@@ -1,0 +1,12 @@
+export interface TokenPrice {
+  address: string;
+  price: number;
+}
+
+export interface Market {
+  mint: string;
+  address: string;
+  name: string;
+  programId: string;
+  deprecated: boolean;
+}
